Move CommentBlock to CommentStyle - the only place where it is currently needed
