+++ src/test/scala/de/heikoseeberger/sbtheader/CommentStyleStec.scala
@@ -17,7 +17,7 @@
 package de.heikoseeberger.sbtheader
 
 import de.heikoseeberger.sbtheader.CommentStyle._
-import org.scalatest.{Matchers, WordSpec}
+import org.scalatest.{ Matchers, WordSpec }
 
 class CommentStyleStec extends WordSpec with Matchers {
